backend/requirements.txt
flask
flask-cors
openai
llama-cpp-python